+++ test/OpenTelemetry.Exporter.Geneva.Tests/EtwCollection.cs
@@ -5,8 +5,8 @@ namespace OpenTelemetry.Exporter.Geneva.Tests;
 
 /// <summary>
 /// Collection definition for ETW tests.
-///
-/// Since ETW is a system-global(not even process-global) resource, these tests shouldn't be run in parallel.
+/// <para/>
+/// Since ETW is a system-global (not even process-global) resource, these tests shouldn't be run in parallel.
 /// </summary>
 [CollectionDefinition(Name, DisableParallelization = true)]
 #pragma warning disable CA1711 // Identifiers should not have incorrect suffix