+++ src/OpenTelemetry.Exporter.Geneva/CHANGELOG.md
@@ -2,6 +2,11 @@
 
 ## Unreleased
 
+* Updated ETW manifest and payload in `EtwDataTransport`
+  Added synthetic `byte[] field` to method signature and `int size` to payload so
+  that runtime-generated .Net ETW manifest matches actual payload
+  ([#4729](https://github.com/open-telemetry/opentelemetry-dotnet-contrib/pull/4729)
+
 * Updated OpenTelemetry core component version(s) to `1.17.0`.
   ([#4773](https://github.com/open-telemetry/opentelemetry-dotnet-contrib/pull/4773))