+++ src/OpenTelemetry.Exporter.Geneva/CHANGELOG.md
@@ -2,6 +2,22 @@
 
 ## Unreleased
 
+## 1.17.0
+
+Released 2026-Jul-17
+
+* Updated OpenTelemetry core component version(s) to `1.17.0`.
+  ([#4773](https://github.com/open-telemetry/opentelemetry-dotnet-contrib/pull/4773))
+
+* Updated ETW manifest and payload in `EtwDataTransport`
+  with synthetic payload so that the runtime-generated .NET
+  ETW manifest matches the actual payload.
+  ([#4729](https://github.com/open-telemetry/opentelemetry-dotnet-contrib/pull/4729)
+
+## 1.16.0
+
+Released 2026-Jul-09
+
 * Updated OpenTelemetry core component version(s) to `1.16.0`.
   ([#4487](https://github.com/open-telemetry/opentelemetry-dotnet-contrib/pull/4487))
 
@@ -13,7 +29,8 @@
   valid names) and freeze the cache on .NET 8+, serialize metric base-128
   strings and HTTP url spans directly into the output buffer instead of via
   temporary arrays/strings.
-  ([#4498](https://github.com/open-telemetry/opentelemetry-dotnet-contrib/pull/4498))
+  ([#4498](https://github.com/open-telemetry/opentelemetry-dotnet-contrib/pull/4498),
+  [#4684](https://github.com/open-telemetry/opentelemetry-dotnet-contrib/pull/4684))
 
 * Assemblies are now digitally signed using cosign.
   ([#4637](https://github.com/open-telemetry/opentelemetry-dotnet-contrib/pull/4637))